[Refactor](Navigation): Moved configs to subfolder.
