+++ lua/config/Navigation/quick_scope.lua
