+++ lua/config/Navigation/telescope.lua
