+++ lua/ds_omega/layers/Navigation/dirbuf.lua
@@ -1,3 +1,3 @@
 local simple_plugin_setup = require('ds_omega.utils').simple_plugin_setup
 
-simple_plugin_setup('dirbuf', 'dirbuf')
+simple_plugin_setup('dirbuf', 'Navigation')
