+++ lua/ds_omega/layers/Navigation/marks.lua
@@ -1,3 +1,3 @@
 local simple_plugin_setup = require('ds_omega.utils').simple_plugin_setup
 
-simple_plugin_setup('marks', 'marks')
+simple_plugin_setup('marks', 'Navigation')
