+++ lua/ds_omega/layers/Navigation/other.lua
@@ -1,3 +1,3 @@
 local simple_plugin_setup = require('ds_omega.utils').simple_plugin_setup
 
-simple_plugin_setup('other-nvim', 'other')
+simple_plugin_setup('other-nvim', 'Navigation.other')
