+++ lua/ds_omega/layers/Navigation/quick_scope.lua
@@ -1,5 +1,5 @@
 local setters = require('utils.setters')
 
-local config = require('config.quick_scope')
+local config = require('config.Navigation.quick_scope')
 
 setters.set_global_variables(config)
