+++ lua/ds_omega/layers/Navigation/telescope.lua
@@ -1,3 +1,3 @@
 local simple_plugin_setup = require('ds_omega.utils').simple_plugin_setup
 
-simple_plugin_setup('telescope', 'telescope')
+simple_plugin_setup('telescope', 'Navigation')
